Lectura comandos

+++ programa_host/src/ringbuffer.c
@@ -3,14 +3,14 @@
 RingBuffer create_ring_buffer() {
     RingBuffer buf;
 
-    buf.data = (uint32_t *) malloc(RING_BUFFER_SIZE * sizeof(uint32_t));
+    buf.data = (uint16_t *) malloc(RING_BUFFER_SIZE * sizeof(uint16_t));
     buf.write_index = 0;
     buf.read_index = 0;
 
     return buf;
 }
 
-int pop_ring_buffer(RingBuffer * buffer, uint32_t * data_ptr) {
+int pop_ring_buffer(RingBuffer * buffer, uint16_t * data_ptr) {
     int next;
 
     if(buffer->read_index == buffer->write_index) {
@@ -27,7 +27,7 @@ int pop_ring_buffer(RingBuffer * buffer, uint32_t * data_ptr) {
     return 0;
 }
 
-int push_ring_buffer(RingBuffer * buffer, uint32_t data_value) {
+int push_ring_buffer(RingBuffer * buffer, uint16_t data_value) {
     int next;
 
     next = buffer->write_index + 1;
